Revert "add own bse::move"
This reverts commit 72b99990b63e7464de360e5cb772387c82dbe045.

+++ c_os/kernel/IOport.h
@@ -37,7 +37,7 @@ public:
     void outb(unsigned char offset, unsigned char val) const {
         asm volatile("outb %0, %1"
                      :
-                     : "a"(val), "Nd"((unsigned short)(address + offset)));
+                     : "a"(val), "Nd"(address + offset));
     }
 
     // Wortweise Ausgabe eines Wertes ueber einen I/O-Port.
@@ -70,7 +70,7 @@ public:
 
         asm volatile("inb %1, %0"
                      : "=a"(ret)
-                     : "Nd"((unsigned short)(address + offset)));
+                     : "Nd"(address + offset));
         return ret;
     }
